Refactor Auth/Login: Move auth logic in core/auth and expose via AuthProvider
The logic, querying and side-effecting for the login and logout, and
accessing the JWT tokens has been moved to `core/auth/*` and is exposed
in the app via the `AuthProvider` in `core/auth/provider.tsx`.

Now, the front-end components are little dumber. They simply get the
user data, login and logout functions from the context and call login,
or logout. They have no knowledge of redirects or interacting with
localStorage etc.

This idea was taken/stolen from the Tanstack Router documentation and
their Authenticated Routes example:
<https://github.com/TanStack/router/tree/main/examples/react/authenticated-routes>